Update build
This commit switches from Maven to Gradle for building the project.
It makes the instructions simpler, by using Gradle plugins:

- the `application` plugin exposes the `run` method directly for the
JVM version, without having to explain to grow the stack size
- the official GraalVM native plugin exposes the `nativeCompile`
and `nativeRun` tasks, respectively for building the native image
and executing it

In addition, this commit switches from using Java 19 to Java 11.
The underlying reason is a [bug](https://github.com/oracle/graal/issues/4594#issuecomment-1330266596)
which prevents building the native image under Linux with AWT,
with more recent versions of Java.

Native configuration has been moved to the conventional
`src/main/resources/META-INF/native-image` location. Although
this isn't required for the JVM version, it simplifies the setup
of the build.